Overhauled site to use templates
Main site theme is now templatized and stored in `_layouts/site.html`

Front-matter added to pages and their templated parts gutted.  Each
page now only includes its content and nothing else.

Added the ability to add headers dynamically to the main layout via
the `header_include` page variable.  This is used as the name of the
file to be included from `_includes/`.